@@ -175,7 +175,7 @@ describe 'MediaObject' do
       it 'renders the transcript viewer' do
         media_object.save
         visit media_object_path(media_object)
-        expect(page.has_css?('div[data-react-class="MediaObjectRamp"]')).to be_truthy
+        expect(page.has_css?('div[id^="MediaObjectRamp-react-component-"]')).to be_truthy
       end
     end
 
@@ -186,7 +186,7 @@ describe 'MediaObject' do
       it 'renders the transcript viewer' do
         media_object.save
         visit media_object_path(media_object)
-        expect(page.has_css?('div[data-react-class="MediaObjectRamp"]')).to be_truthy
+        expect(page.has_css?('div[id^="MediaObjectRamp-react-component-"]')).to be_truthy
       end
     end
 
@@ -199,7 +199,7 @@ describe 'MediaObject' do
         it 'does not render the transcript viewer' do
           media_object.save
           visit media_object_path(media_object)
-          expect(page.has_css?('div[data-react-class="MediaObjectRamp"]')).to be_falsey
+          expect(page.has_css?('div[id^="MediaObjectRamp-react-component-"]')).to be_falsey
         end
       end
       context 'checked out item' do
@@ -208,7 +208,7 @@ describe 'MediaObject' do
         it 'renders the transcript viewer' do
           media_object.save
           visit media_object_path(media_object)
-          expect(page.has_css?('div[data-react-class="MediaObjectRamp"]')).to be_truthy
+          expect(page.has_css?('div[id^="MediaObjectRamp-react-component-"]')).to be_truthy
         end
       end
     end
